Blackjack Backend #6
Created the backbone structure for the Blackjack Minigame, which is the main theme of the demo. How it differs is with 2 different mechanics. 1 is abilities, however those haven't been added into the game yet. The other one has, and that is a player vs player version of traditional blackjack. Instead of playing against a dealer representing a casino, you play against a fellow gambler, betting against each other until one person ends up broke. In order to make this feel like a 1v1, the role of Dealer and Player swap every turn, with the player having the option of setting the betting limit, being able to double down as well as spliting and insurance(adding in a later commit). The dealer, on the other hand, has the main advantage in that if the player busts first, then the dealer wins by default. The renders are flickering, and there are some bugs with the AI that need to be addressed in future commits. This commit also adds a BlackjackAI Class to be used in the future as well as Button classes that will be used.