Unique NPC AI's #43
The BlackjackAI Class allows you to create endless amounts of different blackjack opponents, by changing around the values of aggro, defense, risk, and luck. You can also assign an AI an ability with the BlackjackAI Class, however until #39 is complete this will not be used. Also contains some improvements to the Blackjack code such as implementing the dealer drawing from his own deck instead of a shared one, and when draw is used on an empty deck it reshuffles.