Framework Change
Due to numerous issues discovered with an isometric game created with SDL2, we made the decision to switch over to Unity. This commit represents Sprints 15-18.  Due to a change in not only engines, but also languages, the entire project must be re coded. With this commit, We now have ALL previous features, with the addition of:

- Barrier Physics for floors 0-2
- Functional Main Menu
- Functional Pause Menu
- Dialouge Box
- Easy To Understand Blackjack
- Shop UI
- Floor Keys
- Variable Resolutions
As well as other additional changes/adjustments.